BB-5643: Assetic watch not working - added ScssphpFilterDecorator - check last moditification date to composite input assetic files into one

+++ src/Oro/Bundle/LayoutBundle/Assetic/LayoutResource.php
@@ -4,6 +4,8 @@ namespace Oro\Bundle\LayoutBundle\Assetic;
 
 use Assetic\Factory\Resource\ResourceInterface;
 
+use Psr\Log\LoggerInterface;
+
 use Symfony\Component\Filesystem\Filesystem;
 
 use Oro\Component\Layout\Extension\Theme\Model\Theme;
@@ -23,6 +25,12 @@ class LayoutResource implements ResourceInterface
     /** @var string */
     protected $outputDir;
 
+    /** @var array */
+    protected $mtimeOutputs = [];
+
+    /** @var LoggerInterface */
+    private $logger;
+
     /**
      * @param ThemeManager $themeManager
      * @param Filesystem $filesystem
@@ -36,6 +44,18 @@ class LayoutResource implements ResourceInterface
         $this->themeManager = $themeManager;
         $this->filesystem = $filesystem;
         $this->outputDir = $outputDir;
+    }
+
+    /**
+     * @param LoggerInterface $logger
+     *
+     * @return LayoutResource
+     */
+    public function setLogger(LoggerInterface $logger)
+    {
+        $this->logger = $logger;
+
+        return $this;
     }
 
     /**
@@ -161,17 +181,44 @@ class LayoutResource implements ResourceInterface
         }
         $inputs = array_merge($settingsInputs, $variablesInputs, $restInputs);
 
-        $inputsContent = '';
-        foreach ($inputs as $input) {
-            $inputsContent .= '@import "../'.$input.'"'.";\n";
-        }
-
         $file = realpath($this->outputDir) . '/' . $output . '.'. $extension;
-        $this->filesystem->mkdir(dirname($file), 0777);
-        if (false === @file_put_contents($file, $inputsContent)) {
-            throw new \RuntimeException('Unable to write file ' . $file);
+
+        $mtime = $this->getLastModified($inputs, $extension);
+        if (!array_key_exists($file, $this->mtimeOutputs) || $this->mtimeOutputs[$file] < $mtime) {
+            $inputsContent = '';
+            foreach ($inputs as $input) {
+                $inputsContent .= '@import "../' . $input . '"' . ";\n";
+            }
+
+            $this->filesystem->mkdir(dirname($file), 0777);
+            if (false === @file_put_contents($file, $inputsContent)) {
+                throw new \RuntimeException('Unable to write file ' . $file);
+            }
+
+            $this->mtimeOutputs[$file] = $mtime;
         }
 
         return $file;
+    }
+
+    /**
+     * @param array  $inputs
+     * @param string $extension
+     *
+     * @return int|mixed
+     */
+    private function getLastModified(array $inputs, $extension)
+    {
+        $mtime = 0;
+        foreach ($inputs as $input) {
+            $file = realpath($this->outputDir) . '/' . $input;
+            if (file_exists($file)) {
+                $mtime = max($mtime, filemtime($file));
+            } else {
+                $this->logger->debug(sprintf('Could not find file %s, declared in assets.yml.', $input));
+            }
+        }
+
+        return $mtime;
     }
 }
